6/p: remove split of executeDeploy to deploy
This also makes it not update its own deploy status
and use runWithLog; it's both simpler and supplies the error handling we want
