+++ src/common/delivery/deploy/executeUndeploy.ts
@@ -20,7 +20,8 @@ import { DeployableArtifact } from "../../../spi/artifact/ArtifactStore";
 import { TargetInfo } from "../../../spi/deploy/Deployment";
 import { ProgressLog } from "../../../spi/log/ProgressLog";
 import { GoalExecutor } from "../goals/goalExecution";
-import { DeploySpec, ExecuteWithLog, runWithLog, RunWithLogContext } from "./runWithLog";
+import { ExecuteWithLog, runWithLog, RunWithLogContext } from "./runWithLog";
+import { DeploySpec } from "./executeDeploy";
 
 export function undeployArtifactWithLogs<T extends TargetInfo>(spec: DeploySpec<T>): GoalExecutor {
     return runWithLog(executeUndeployArtifact(spec), spec.deployer.logInterpreter);
