+++ src/software-delivery-machine/blueprint/deploy/cloudFoundryDeploy.ts
@@ -19,7 +19,6 @@ import {
     EnvironmentCloudFoundryTarget,
 } from "../../../common/delivery/deploy/pcf/CloudFoundryTarget";
 import { CommandLineCloudFoundryDeployer } from "../../../common/delivery/deploy/pcf/CommandLineCloudFoundryDeployer";
-import { DeploySpec } from "../../../common/delivery/deploy/runWithLog";
 import {
     ProductionDeploymentGoal,
     ProductionEndpointGoal,
@@ -31,7 +30,7 @@ import { ProjectLoader } from "../../../common/repo/ProjectLoader";
 import { setDeployEnablement } from "../../../handlers/commands/SetDeployEnablement";
 import { ArtifactStore } from "../../../spi/artifact/ArtifactStore";
 import { AddCloudFoundryManifestMarker } from "../../commands/editors/pcf/addCloudFoundryManifest";
-import { DefaultArtifactStore } from "../artifactStore";
+import { DeploySpec } from "../../../common/delivery/deploy/executeDeploy";
 
 /**
  * Deploy everything to the same Cloud Foundry space
